feat: implement unified email hub for admin
- Created EmailHub.tsx to consolidate Compose, Broadcasts, Templates, and Stats.
- Updated send-custom-email edge function with recipient targeting and HTML template wrapping.
- Migrated Raffle Announcement logic to the new hub.
- Removed redundant EmailCenter, EmailPreview, and NewsletterSender components.
- Updated Admin Dashboard and Sidebar navigation.
